+++ test/ui/screens/home_screen/widgets/posts_list/index_test.dart
@@ -42,7 +42,8 @@ void main() {
       ),
     );
 
-    await tester.pumpAndSettle();
+    await tester.pump(const Duration(seconds: 3));
+
     expect(find.byType(PostsLoadingEmptyContainer), findsNothing);
     expect(find.byType(PostListItem), findsWidgets);
     expect(find.byType(CustomScrollView), findsOneWidget);
@@ -91,7 +92,7 @@ void main() {
       ),
     );
 
-    await tester.pumpAndSettle();
+    await tester.pump(const Duration(seconds: 3));
     expect(find.byType(ErrorPostMessage), findsOneWidget);
     expect(find.byType(ErrorPost), findsOneWidget);
     expect(find.byIcon(MooncakeIcons.syncing), findsOneWidget);