+++ test/ui/screens/post_details_screen/widgets/post_details_main_content/index_test.dart
@@ -10,27 +10,27 @@ import 'package:mooncake/entities/entities.dart';
 import 'package:mooncake/ui/widgets/post_content/widgets/export.dart';
 import 'package:mooncake/ui/screens/post_details_screen/widgets/post_details_main_content/widgets/export.dart';
 import 'package:mooncake/ui/screens/post_details_screen/widgets/post_details_main_content/widgets/post_comments_list/widgets/export.dart';
+import 'package:mooncake/ui/screens/post_details_screen/widgets/post_details_main_content/widgets/post_details_reactions_list/widgets/reactions_list/widgets/export.dart';
 import 'package:mooncake/ui/screens/post_details_screen/widgets/post_details_main_content/widgets/post_details_reactions_list/widgets/export.dart';
 
 void main() {
+  MockPostDetailsBloc mockPostDetailsBloc = MockPostDetailsBloc();
+  MockAccountBloc mockAccountBloc = MockAccountBloc();
+  MockNavigatorBloc mockNavigatorBloc = MockNavigatorBloc();
+  MooncakeAccount userAccount = MooncakeAccount(
+    profilePicUri: "https://example.com/avatar.png",
+    moniker: "john-doe",
+    cosmosAccount: CosmosAccount(
+      accountNumber: 153,
+      address: "desmos1ew60ztvqxlf5kjjyyzxf7hummlwdadgesu3725",
+      coins: [
+        StdCoin(amount: "10000", denom: "udaric"),
+      ],
+      sequence: 45,
+    ),
+  );
   testWidgets('PostDetailsMainContent: Displays correctly',
       (WidgetTester tester) async {
-    MockPostDetailsBloc mockPostDetailsBloc = MockPostDetailsBloc();
-    MockAccountBloc mockAccountBloc = MockAccountBloc();
-    MockNavigatorBloc mockNavigatorBloc = MockNavigatorBloc();
-    MooncakeAccount userAccount = MooncakeAccount(
-      profilePicUri: "https://example.com/avatar.png",
-      moniker: "john-doe",
-      cosmosAccount: CosmosAccount(
-        accountNumber: 153,
-        address: "desmos1ew60ztvqxlf5kjjyyzxf7hummlwdadgesu3725",
-        coins: [
-          StdCoin(amount: "10000", denom: "udaric"),
-        ],
-        sequence: 45,
-      ),
-    );
-
     when(mockPostDetailsBloc.state).thenReturn(PostDetailsLoaded.first(
       user: userAccount,
       post: testPost,
@@ -67,6 +67,11 @@ void main() {
     expect(find.byType(InkWell), findsWidgets);
     expect(find.byType(PostReactionItem), findsNothing);
     expect(find.byType(PostDetailsReactionsList), findsNothing);
+    expect(find.byType(PostDetailsBottomBar), findsOneWidget);
+    expect(find.byType(PostCommentAction), findsWidgets);
+    expect(find.byType(PostAddReactionAction), findsWidgets);
+    expect(find.byType(PostLikeAction), findsNothing);
+    expect(find.byType(Tab), findsNWidgets(2));
     expect(
       find.descendant(
         of: find.byType(PostCommentsList),
@@ -74,8 +79,6 @@ void main() {
       ),
       findsWidgets,
     );
-
-    // expect(find.byIcon(MooncakeIcons.poll), findsOneWidget);
 
     await tester.tap(find
         .descendant(
@@ -85,5 +88,102 @@ void main() {
         .first);
     await tester.pumpAndSettle();
     expect(verify(mockNavigatorBloc.add(any)).callCount, 1);
+  });
+
+  testWidgets('PostDetailsMainContent: Displays empty reactions correctly',
+      (WidgetTester tester) async {
+    when(mockPostDetailsBloc.state).thenReturn(PostDetailsLoaded.first(
+      user: userAccount,
+      post: testPost,
+      comments: testPosts,
+    ));
+
+    when(mockAccountBloc.state).thenReturn(LoggedIn.initial(userAccount));
+
+    await tester.pumpWidget(
+      makeTestableWidget(
+        child: MultiBlocProvider(
+          providers: [
+            BlocProvider<PostDetailsBloc>(
+              create: (_) => mockPostDetailsBloc,
+            ),
+            BlocProvider<AccountBloc>(
+              create: (_) => mockAccountBloc,
+            ),
+            BlocProvider<NavigatorBloc>(
+              create: (_) => mockNavigatorBloc,
+            ),
+          ],
+          child: PostDetailsMainContent(),
+        ),
+      ),
+    );
+
+    await tester.pumpAndSettle();
+    await tester.tap(find.byType(Tab).last);
+    await tester.pumpAndSettle();
+
+    expect(find.byType(PostDetailsReactionsList), findsOneWidget);
+    expect(find.byType(EmptyReactions), findsOneWidget);
+    expect(find.text("noReactionsYet"), findsOneWidget);
+  });
+
+  testWidgets('PostDetailsMainContent: Displays reactions correctly',
+      (WidgetTester tester) async {
+    List<Reaction> reactionTest = [
+      Reaction(user: userAccount, value: "123", code: "123"),
+      Reaction(user: userAccount, value: "123", code: "123"),
+      Reaction(user: userAccount, value: "123", code: "123"),
+      Reaction(user: userAccount, value: "3", code: "123"),
+    ];
+
+    when(mockPostDetailsBloc.state).thenReturn(PostDetailsLoaded.first(
+      user: userAccount,
+      post: testPost.copyWith(reactions: reactionTest),
+      comments: testPosts,
+    ));
+
+    when(mockAccountBloc.state).thenReturn(LoggedIn.initial(userAccount));
+
+    await tester.pumpWidget(
+      makeTestableWidget(
+        child: MultiBlocProvider(
+          providers: [
+            BlocProvider<PostDetailsBloc>(
+              create: (_) => mockPostDetailsBloc,
+            ),
+            BlocProvider<AccountBloc>(
+              create: (_) => mockAccountBloc,
+            ),
+            BlocProvider<NavigatorBloc>(
+              create: (_) => mockNavigatorBloc,
+            ),
+          ],
+          child: PostDetailsMainContent(),
+        ),
+      ),
+    );
+
+    await tester.pumpAndSettle();
+    await tester.tap(find.byType(Tab).last);
+    await tester.pumpAndSettle();
+
+    expect(find.byType(PostDetailsReactionsList), findsOneWidget);
+    expect(find.byType(EmptyReactions), findsNothing);
+    expect(find.byType(ReactionFilterItem), findsWidgets);
+    expect(find.byType(PostReactionItem), findsNWidgets(4));
+    expect(find.text("noReactionsYet"), findsNothing);
+
+    await tester.tap(find.text("3 1"));
+    await tester.pumpAndSettle();
+    expect(find.byType(PostReactionItem), findsNWidgets(1));
+    expect(find.byType(AccountAvatar), findsWidgets);
+
+    await tester.tap(find.byType(AccountAvatar).last);
+    await tester.pumpAndSettle();
+    expect(
+        verify(mockNavigatorBloc.add(NavigateToUserDetails(userAccount)))
+            .callCount,
+        1);
   });
 }